% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_simulation.R
\name{simulate_count_regular}
\alias{simulate_count_regular}
\title{Simulate a count matrix based on input model directly}
\usage{
simulate_count_regular(
  model_params,
  n_cell_new,
  cell_type_prop = 1,
  sim_method = c("copula", "ind"),
  cell_sample = FALSE
)
}
\arguments{
\item{model_params}{A list with the same length as \code{cell_type_prop} that contains
the fitted model as each of its element (can be either the copula
model or the (w/o copula) model).}

\item{n_cell_new}{An integer value that indicates the number of cells to generate.}

\item{cell_type_prop}{The cell type proportion in the simulated count matrix.}

\item{sim_method}{Specification of the type of model for data simulation.
Default value is 'copula', which selects the copula model.
'ind' will select the (w/o copula) model.}

\item{cell_sample}{Logical, whether cells for each cell type should be sampled from a
multinomial distribution or follows the exact same proportion as
specified in \code{cell_type_prop}.}
}
\description{
Simulate a count matrix based on input model directly
}
