% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_simulation.R
\name{simulate_count_scDesign2}
\alias{simulate_count_scDesign2}
\title{Simulate a count matrix for experimental design}
\usage{
simulate_count_scDesign2(
  model_params,
  n_cell_new,
  cell_type_prop = 1,
  total_count_new = NULL,
  total_count_old = NULL,
  n_cell_old = NULL,
  sim_method = c("copula", "ind"),
  reseq_method = c("mean_scale", "multinomial"),
  cell_sample = FALSE
)
}
\arguments{
\item{model_params}{A list with the same length as \code{cell_type_prop} that contains
the fitted model as each of its element (can be either the copula
model or the (w/o copula) model).}

\item{n_cell_new}{The total number of cells in the simulated count matrix.}

\item{cell_type_prop}{The cell type proportion in the simulated count matrix.}

\item{total_count_new}{The (expected) total number of reads or UMIs in the simulated count
matrix.}

\item{total_count_old}{The total number of reads or UMIs in the original count matrix where
\code{model_params} was fitted.}

\item{n_cell_old}{The The total number of cells in the original count matrix where
\code{model_params} was fitted.}

\item{sim_method}{Specification of the type of model for data simulation.
Default value is 'copula', which selects the copula model.
'ind' will select the (w/o copula) model.}

\item{reseq_method}{Specification of how the new count matrix should be derived under the
new sequencing depth.
Default is 'mean_scale', which scales the original parameters and
then simulate new data.
'multinomial' will do a resampling. It ensures that the simulated
count matrix has the exact total number of reads as specified in
\code{total_count_new}.}

\item{cell_sample}{Logical, whether cells for each cell type should be sampled from a
multinomial distribution or follows the exact same proportion as
specified in \code{cell_type_prop}.}
}
\value{
A matrix of shape p by n that contains the simulated count values. p is derived from
\code{model_params}.
}
\description{
Simulate a count matrix for experimental design
}
